requirements.txt
torch
torchvision
numpy
pillow
pyyaml
matplotlib
tqdm
optuna
opencv-python
